fix ssg of prev api
